--- a/src/crypto/CMakeLists.txt
+++ b/src/crypto/CMakeLists.txt
@@ -3,6 +3,8 @@ set(crypto_plugin_dir ${CEPH_INSTALL_PKGLIBDIR}/crypto)
 
 add_subdirectory(openssl)
 
+add_subdirectory(ipsec_mb)
+
 if(HAVE_INTEL AND HAVE_NASM_X64_AVX2 AND (NOT APPLE))
   add_subdirectory(isa-l)
 endif()