--- a/src/crypto/crypto_accel.h
+++ b/src/crypto/crypto_accel.h
@@ -20,6 +20,18 @@
 class CryptoAccel;
 typedef std::shared_ptr<CryptoAccel> CryptoAccelRef;
 
+class CryptoEngine {
+ public:
+  static const int AES_256_KEYSIZE = 256 / 8;
+  unsigned char m_key[AES_256_KEYSIZE];
+
+  CryptoEngine() {}
+  CryptoEngine(const unsigned char(&key)[AES_256_KEYSIZE]) {
+    memcpy(m_key, key, sizeof(key));
+  }
+  virtual ~CryptoEngine() {}
+};
+
 class CryptoAccel {
  public:
   CryptoAccel() {}
@@ -33,5 +45,28 @@ class CryptoAccel {
   virtual bool cbc_decrypt(unsigned char* out, const unsigned char* in, size_t size,
                    const unsigned char (&iv)[AES_256_IVSIZE],
                    const unsigned char (&key)[AES_256_KEYSIZE]) = 0;
+
+  /*
+  * Operate on multiple buffers in parallel and flush at the end, aim to promote performance.
+  * Need to configure ipsec as the accelerator, refer to:
+  *  - option: plugin_crypto_accelerator
+  *  - value: crypto_ipsec
+  *
+  * Others like openssl/isa-l/QAT don't have multi-buffer functionality, so switch to traditional interface
+  */
+  virtual CryptoEngine* get_engine(const unsigned char(&key)[AES_256_KEYSIZE]) { return new CryptoEngine(key); }
+  virtual void put_engine(CryptoEngine* eng) { delete eng; }
+
+  virtual bool flush(CryptoEngine* engine) { return true; }
+  virtual bool cbc_encrypt_mb(CryptoEngine* engine, unsigned char* out, const unsigned char* in, size_t size,
+    const unsigned char(&iv)[AES_256_IVSIZE]) {
+    if (!engine) return false;
+    return cbc_encrypt(out, in, size, iv, engine->m_key);
+  };
+  virtual bool cbc_decrypt_mb(CryptoEngine* engine, unsigned char* out, const unsigned char* in, size_t size,
+    const unsigned char(&iv)[AES_256_IVSIZE]) {
+    if (!engine) return false;
+    return cbc_decrypt(out, in, size, iv, engine->m_key);
+  };
 };
 #endif