Currently we have 3 types of crypto accelerators: isa-l/openssl/QAT, this PR is trying to introduce the fourth crypto accelerator: ipsec_mb. Like isa-l and openssl, ipsec_mb is just a software accelerator. The big advantage is: ipsec_mb module can leverage latest instruction set(VAES) of processor, this can significantly improve encryption/decryption speed.

Performance comparison (AES-CBC per CPU core):

    [encryption]
    - crypto/ipsec_mb with VAES: ~12GB/s
    - crypto/isa-l: ~1.2GB/s
    - crypto/openssl: ~1.2GB/s

    [decryption]
    - crypto/ipsec_mb with VAES: ~14GB/s
    - crypto/isa-l: ~6GB/s
    - crypto/openssl: ~6GB/s

So ipsec_mb has about 10 times performance gain comparing with existing accelerators (isa-l and openssl) in encryption, and more than 2 times in decryption.

* Measured via test_crypto_plugins.cc in this PR, may vary on different platforms

Regarding the performance gain in RGW, it depends on platforms. Using minio/warp to do the test, you may see ~7% gain in Xeon(R) Gold 6139. But you may can only see 1% or even cannot see obvious performance gain in some platforms. This may be caused by: the frequency of CPU core which running avx512 will be slightly drop down, e.g. 3.2GHz->3.1GHz. The ideal use model would be let this plugin running in a seperate thread(cpu core), so the frequency of other CPU core would remain un-impacted.

Signed-off-by: Liang Fang <liang.a.fang@intel.com>